Deprecate netty server query
