+++ presto-main/etc/catalog/pinot.properties
@@ -19,11 +19,8 @@ pinot.controller-urls=localhost:9000
 ## Pinot is behind a Proxy
 #pinot.proxy-enabled=true
 
-## Requires Pinot version >= 0.6.0.
-## Enable Query Pinot streaming server for segment queries via gRPC
-# pinot.use-streaming-for-segment-queries=true
-## gRPC host and port, required when streaming query is enabled
-# pinot.grpc-host=broker.cogi.presto-demo.daniel.startree-dev.cloud
+## Pinot Proxy gRPC host and port, required when Pinot Proxy is enabled.
+# pinot.grpc-host=proxy.grpc.host
 # pinot.grpc-port=8094
 
 # Specify keystore file and keystore password for two-way tls
