+++ presto-pinot-toolkit/src/main/java/com/facebook/presto/pinot/PinotSessionProperties.java
@@ -17,14 +17,12 @@ package com.facebook.presto.pinot;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.session.PropertyMetadata;
 import com.google.common.collect.ImmutableList;
-import io.airlift.units.Duration;
 
 import javax.inject.Inject;
 
 import java.util.List;
 
 import static com.facebook.presto.common.type.IntegerType.INTEGER;
-import static com.facebook.presto.common.type.VarcharType.createUnboundedVarcharType;
 import static com.facebook.presto.spi.session.PropertyMetadata.booleanProperty;
 import static com.facebook.presto.spi.session.PropertyMetadata.integerProperty;
 import static com.facebook.presto.spi.session.PropertyMetadata.stringProperty;
@@ -32,10 +30,8 @@ import static com.google.common.base.Preconditions.checkArgument;
 
 public class PinotSessionProperties
 {
-    public static final String CONNECTION_TIMEOUT = "connection_timeout";
     public static final String FORBID_BROKER_QUERIES = "forbid_broker_queries";
     public static final String ATTEMPT_BROKER_QUERIES = "attempt_broker_queries";
-    public static final String IGNORE_EMPTY_RESPONSES = "ignore_empty_responses";
     public static final String RETRY_COUNT = "retry_count";
     public static final String MARK_DATA_FETCH_EXCEPTIONS_AS_RETRIABLE = "mark_data_fetch_exceptions_as_retriable";
     public static final String USE_DATE_TRUNC = "use_date_trunc";
@@ -73,16 +69,6 @@ public class PinotSessionProperties
     public static boolean isAttemptBrokerQueries(ConnectorSession session)
     {
         return session.getProperty(ATTEMPT_BROKER_QUERIES, Boolean.class);
-    }
-
-    public static Duration getConnectionTimeout(ConnectorSession session)
-    {
-        return session.getProperty(CONNECTION_TIMEOUT, Duration.class);
-    }
-
-    public static boolean isIgnoreEmptyResponses(ConnectorSession session)
-    {
-        return session.getProperty(IGNORE_EMPTY_RESPONSES, Boolean.class);
     }
 
     public static int getPinotRetryCount(ConnectorSession session)
@@ -169,11 +155,6 @@ public class PinotSessionProperties
                         "Attempt broker queries",
                         pinotConfig.isAttemptBrokerQueries(),
                         false),
-                booleanProperty(
-                        IGNORE_EMPTY_RESPONSES,
-                        "Ignore empty or missing pinot server responses",
-                        pinotConfig.isIgnoreEmptyResponses(),
-                        false),
                 integerProperty(
                         RETRY_COUNT,
                         "Retry count for retriable pinot data fetch calls",
@@ -239,15 +220,6 @@ public class PinotSessionProperties
                         "Push down expressions in projection to Pinot broker",
                         pinotConfig.isPushdownProjectExpressions(),
                         false),
-                new PropertyMetadata<>(
-                        CONNECTION_TIMEOUT,
-                        "Connection Timeout to talk to Pinot servers",
-                        createUnboundedVarcharType(),
-                        Duration.class,
-                        pinotConfig.getConnectionTimeout(),
-                        false,
-                        value -> Duration.valueOf((String) value),
-                        Duration::toString),
                 new PropertyMetadata<>(
                         NUM_SEGMENTS_PER_SPLIT,
                         "Number of segments of the same host per split",
