+++ presto-pinot-toolkit/src/test/java/com/facebook/presto/pinot/TestPinotConfig.java
@@ -33,22 +33,15 @@ public class TestPinotConfig
                         .setExtraHttpHeaders("")
                         .setExtraGrpcMetadata("")
                         .setControllerUrls("")
-                        .setIdleTimeout(new Duration(5, TimeUnit.MINUTES))
                         .setLimitLargeForSegment(PinotConfig.DEFAULT_LIMIT_LARGE_FOR_SEGMENT)
                         .setTopNLarge(PinotConfig.DEFAULT_TOPN_LARGE)
-                        .setMaxBacklogPerServer(PinotConfig.DEFAULT_MAX_BACKLOG_PER_SERVER)
-                        .setMaxConnectionsPerServer(PinotConfig.DEFAULT_MAX_CONNECTIONS_PER_SERVER)
-                        .setMinConnectionsPerServer(PinotConfig.DEFAULT_MIN_CONNECTIONS_PER_SERVER)
-                        .setThreadPoolSize(PinotConfig.DEFAULT_THREAD_POOL_SIZE)
                         .setEstimatedSizeInBytesForNonNumericColumn(20)
-                        .setConnectionTimeout(new Duration(1, TimeUnit.MINUTES))
                         .setControllerRestService(null)
                         .setServiceHeaderParam("RPC-Service")
                         .setCallerHeaderValue("presto")
                         .setCallerHeaderParam("RPC-Caller")
                         .setOverrideDistinctCountFunction("distinctCount")
                         .setMetadataCacheExpiry(new Duration(2, TimeUnit.MINUTES))
-                        .setAllowMultipleAggregations(true)
                         .setInferDateTypeInSchema(true)
                         .setInferTimestampTypeInSchema(true)
                         .setForbidBrokerQueries(false)
@@ -68,11 +61,9 @@ public class TestPinotConfig
                         .setMarkDataFetchExceptionsAsRetriable(true)
                         .setPushdownTopNBrokerQueries(true)
                         .setPushdownProjectExpressions(true)
-                        .setIgnoreEmptyResponses(false)
                         .setUseDateTrunc(false)
                         .setForbidSegmentQueries(false)
                         .setAttemptBrokerQueries(false)
-                        .setUseStreamingForSegmentQueries(true)
                         .setStreamingServerGrpcMaxInboundMessageBytes(PinotConfig.DEFAULT_STREAMING_SERVER_GRPC_MAX_INBOUND_MESSAGE_BYTES)
                         .setNonAggregateLimitForBrokerQueries(PinotConfig.DEFAULT_NON_AGGREGATE_LIMIT_FOR_BROKER_QUERIES)
                         .setUseDateTrunc(false)
@@ -92,19 +83,12 @@ public class TestPinotConfig
                 .put("pinot.extra-grpc-metadata", "k1:v1")
                 .put("pinot.controller-rest-service", "pinot-controller-service")
                 .put("pinot.controller-urls", "host1:1111,host2:1111")
-                .put("pinot.idle-timeout", "1h")
                 .put("pinot.topn-large", "1000")
-                .put("pinot.max-backlog-per-server", "15")
-                .put("pinot.max-connections-per-server", "10")
-                .put("pinot.min-connections-per-server", "1")
-                .put("pinot.thread-pool-size", "100")
                 .put("pinot.estimated-size-in-bytes-for-non-numeric-column", "30")
-                .put("pinot.connection-timeout", "8m")
                 .put("pinot.metadata-expiry", "1m")
                 .put("pinot.caller-header-value", "myCaller")
                 .put("pinot.caller-header-param", "myParam")
                 .put("pinot.service-header-param", "myServiceHeader")
-                .put("pinot.allow-multiple-aggregations", "false")
                 .put("pinot.infer-date-type-in-schema", "false")
                 .put("pinot.infer-timestamp-type-in-schema", "false")
                 .put("pinot.forbid-broker-queries", "true")
@@ -113,7 +97,6 @@ public class TestPinotConfig
                 .put("pinot.grpc-port", "8224")
                 .put("pinot.proxy-enabled", "true")
                 .put("pinot.num-segments-per-split", "2")
-                .put("pinot.ignore-empty-responses", "true")
                 .put("pinot.fetch-retry-count", "3")
                 .put("pinot.mark-data-fetch-exceptions-as-retriable", "false")
                 .put("pinot.non-aggregate-limit-for-broker-queries", "10")
@@ -123,7 +106,6 @@ public class TestPinotConfig
                 .put("pinot.pushdown-project-expressions", "false")
                 .put("pinot.forbid-segment-queries", "true")
                 .put("pinot.attempt-broker-queries", "true")
-                .put("pinot.use-streaming-for-segment-queries", "false")
                 .put("pinot.streaming-server-grpc-max-inbound-message-bytes", "65536")
                 .put("pinot.secure-connection", "true")
                 .put("pinot.override-distinct-count-function", "distinctCountBitmap")
@@ -148,27 +130,19 @@ public class TestPinotConfig
                 .setControllerUrls("host1:1111,host2:1111")
                 .setGrpcHost("localhost")
                 .setGrpcPort(8224)
-                .setIdleTimeout(new Duration(1, TimeUnit.HOURS))
                 .setLimitLargeForSegment(100000)
                 .setTopNLarge(1000)
-                .setMaxBacklogPerServer(15)
-                .setMaxConnectionsPerServer(10)
-                .setMinConnectionsPerServer(1)
-                .setThreadPoolSize(100)
                 .setEstimatedSizeInBytesForNonNumericColumn(30)
-                .setConnectionTimeout(new Duration(8, TimeUnit.MINUTES))
                 .setServiceHeaderParam("myServiceHeader")
                 .setCallerHeaderValue("myCaller")
                 .setCallerHeaderParam("myParam")
                 .setOverrideDistinctCountFunction("distinctCountBitmap")
                 .setMetadataCacheExpiry(new Duration(1, TimeUnit.MINUTES))
-                .setAllowMultipleAggregations(false)
                 .setInferDateTypeInSchema(false)
                 .setInferTimestampTypeInSchema(false)
                 .setForbidBrokerQueries(true)
                 .setRestProxyServiceForQuery("pinot-rest-proxy-service")
                 .setNumSegmentsPerSplit(2)
-                .setIgnoreEmptyResponses(true)
                 .setFetchRetryCount(3)
                 .setMarkDataFetchExceptionsAsRetriable(false)
                 .setNonAggregateLimitForBrokerQueries(10)
@@ -177,7 +151,6 @@ public class TestPinotConfig
                 .setPushdownProjectExpressions(false)
                 .setForbidSegmentQueries(true)
                 .setAttemptBrokerQueries(true)
-                .setUseStreamingForSegmentQueries(false)
                 .setStreamingServerGrpcMaxInboundMessageBytes(65536)
                 .setUseDateTrunc(true)
                 .setUseProxy(true)
